Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
alembic>=1.14.0
redis>=5.0
httpx>=0.27.0
python-multipart>=0.0.9
aiofiles>=24.0
faiss-cpu>=1.8.0
numpy>=1.26.0
google-generativeai>=0.8.0
sentence-transformers>=3.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
